test: update structural tests for extracted components
